Solve day 24 part 1

+++ kotlin/common/src/main/kotlin/common/geom/Points.kt
@@ -20,7 +20,7 @@ val origin2D = Point2D(0, 0)
 fun p2(x: Int, y: Int): Point2D =
     Point2D(x, y)
 
-data class Point3D(val x: Int, val y: Int, val z: Int)
+data class Point3D(val x: Long, val y: Long, val z: Long)
 
 operator fun Point3D.unaryMinus(): Point3D =
     copy(x = -x, y = -y, z = -z)
@@ -34,7 +34,7 @@ operator fun Point3D.minus(other: Point3D): Point3D =
 operator fun Int.times(point3D: Point3D): Point3D =
     point3D.copy(x = this * point3D.x, y = this *  point3D.y, z = this * point3D.z)
 
-val origin3D = Point3D(0, 0, 0)
+val origin3D = Point3D(0L, 0L, 0L)
 
-fun p3(x: Int, y: Int, z:Int): Point3D =
+fun p3(x: Long, y: Long, z:Long): Point3D =
     Point3D(x, y, z)
